Reference

Reference

HSL MA57 and MA97 can be used for the solution of symmetric, possibly indefinite, linear systems. They are often used for the solution of saddle-point systems, i.e., systems of the form

\[\begin{bmatrix} H & A^T \\ A & \end{bmatrix} \begin{bmatrix} x \\ y \end{bmatrix} = \begin{bmatrix} b \\ c \end{bmatrix}.\]

Special cases occur when $H$ is the identity matrix and either $b = 0$ or $c = 0$, which correspond to the least-squares problem

\[\min_y \ \|A^T y - b\|\]

and to the least-norm problem

\[\min_x \ \|x\| \ \text{subject to } Ax = c\]

respectively.

HSL.Ma97 — Method.

Instantiate and perform symbolic analysis on a sparse Julia matrix

M = Ma97(A; kwargs...)

Instantiate an object of type Ma97 and perform the symbolic analysis on a sparse Julia matrix.

Input arguments

A::SparseMatrixCSC{T<:Ma97Data,Int}: input matrix. The lower triangle will be extracted.

Keyword arguments

All keyword arguments are passed directly to ma97_csc().

HSL.Ma97Exception — Type.

Exception type raised in case of error.

HSL.Ma97_Control — Type.

Main control type for MA97.

Ma97_Control(; kwargs...)

Keyword arguments:

print_level::Int: integer controling the verbosit level. Accepted values are:

<0: no printing

0: errors and warnings only (default)

1: errors, warnings and basic diagnostics

2: errors, warning and full diagnostics

unit_diagnostics::Int: Fortran file unit for diagnostics (default: 6)

unit_error::Int: Fortran file unit for errors (default: 6)

unit_warning::Int: Fortran file unit for warnings (default: 6)

HSL.Ma97_Info — Type.

Main info type for MA97

info = Ma97_Info{T <: Ma97Real}()

An info variable is used to collect statistics on the analysis, factorization and solve.

HSL.ma97_coord — Function.

Instantiate and perform symbolic analysis using coordinate arrays

M = ma97_coord(n, cols, rows, nzval; kwargs...)

Instantiate an object of type Ma97 and perform the symbolic analysis on a matrix described in sparse coordinate format.

Input arguments

n::Int: the matrix size

cols::Vector{T<:Integer}: array of column indices for the lower triangle

rows::Vector{T<:Integer}: array of row indices for the lower triangle

nzval::Vector{T<:Ma97Data}: array of values for the lower triangle

Keyword arguments

All keyword arguments are passed directly to the Ma97_Control constructor.

HSL.ma97_csc — Function.

Instantiate and perform symbolic analysis using CSC arrays

M = ma97_csc(n, colptr, rowval, nzval; kwargs...)

Instantiate an object of type Ma97 and perform the symbolic analysis on a matrix described in sparse CSC format.

Input arguments

n::Int: the matrix size

colptr::Vector{T<:Integer}: CSC colptr array for the lower triangle

rowval::Vector{T<:Integer}: CSC rowval array for the lower triangle

nzval::Vector{T<:Ma97Data}: CSC nzval array for the lower triangle

Keyword arguments

All keyword arguments are passed directly to the Ma97_Control constructor.

HSL.ma97_enquire — Function.

ma97enquire: see the documentation for `ma97inquire`.

HSL.ma97_factorise — Function.

ma97factorise: see the documentation for `ma97factorize`.

HSL.ma97_factorise! — Function.

ma97factorise!: see the documentation for `ma97factorize!`.

HSL.ma97_factorize — Function.

Combined Analysis and factorization

M = ma97_factorize(A; kwargs...)

Convenience method that combines the symbolic analysis and numerical factorization phases. An MA97 instance is returned, that can subsequently be passed to other functions, e.g., ma97_solve().

Input Arguments

A::SparseMatrixCSC{T<:Ma97Data,Int}: Julia sparse matrix

Keyword Arguments

matrix_type::Symbol=:real_indef: indicates the matrix type. Accepted values are

:real_spd for a real symmetric and positive definite matrix

:real_indef for a real symmetric and indefinite matrix.

HSL.ma97_factorize! — Function.

Perform numerical factorization.

ma97_factorize!(ma97; kwargs...)

The symbolic analysis must have been performed and must have succeeded.

Input Arguments

ma97::Ma97{T<:Ma97Data}:: an Ma97 structure for which the analysis has been performed

Keyword Arguments

matrix_type::Symbol=:real_indef: indicates the matrix type. Accepted values are

:real_spd for a real symmetric and positive definite matrix

:real_indef for a real symmetric and indefinite matrix.

HSL.ma97_inquire — Function.

Inquire about a factorization or solve

ma97_inquire(ma97; kwargs...)

Obtain information on the pivots after a successful factorization or solve.

Input Arguments

ma97::Ma97{T<:Ma97Data}: an Ma97 structure for which the analysis and factorization have been performed

Keyword arguments

matrix_type::Symbol=:real_indef: indicates the matrix type. Accepted values are

:real_spd for a real symmetric and positive definite matrix

:real_indef for a real symmetric and indefinite matrix.

Return values

An inquiry on a real or complex indefinite matrix returns two vectors:

piv_order: contains the pivot sequence; a negative value indicates that the corresponding variable is part of a 2x2 pivot,

d: a 2 by n array whose first row contains the diagonal of D⁻¹ in the factorization, and whose nonzeros in the second row contain the off-diagonals.

An inquiry on a positive definite matrix returns one vector with the pivot values.

HSL.ma97_least_squares — Method.

Solve least-squares problem

ma97_least_squares(A, b)

Solve the least-squares problem

minimize ‖Ax - b‖₂

where A has shape m-by-n with m > n, by solving the saddle-point system

[ I   A ] [ r ]   [ b ]
[ A'    ] [ x ] = [ 0 ].

Input arguments

A::SparseMatrixCSC{T<:Ma97Data,Int}: input matrix of shape m-by-n with m > n. A full matrix will be converted to sparse.

b::Vector{T}: right-hand side vector

Return value

x::Vector{T}: solution vector.

HSL.ma97_min_norm — Method.

Solve a minimum-norm problem

ma97_min_norm(A, b)

solves

minimize ‖x‖₂  subject to Ax=b,

where A has shape m-by-n with m < n, by solving the saddle-point system

[ I  A' ] [ x ]   [ 0 ]
[ A     ] [ y ] = [ b ].

Input arguments

A::SparseMatrixCSC{T<:Ma97Data,Int}: input matrix of shape m-by-n with m < n. A full matrix will be converted to sparse.

b::Vector{T}: right-hand side vector

Return value

x::Vector{T}: solution vector.

HSL.ma97_solve — Function.

System solve

Solve after factorization

ma97_solve(ma97, b; kwargs...)

Input arguments

ma97::Ma97{T<:Ma97Data}: an Ma97 structure for which the analysis and factorization have been performed

b::Array{T}: vector of array of right-hand sides. Note that b will be overwritten. To solve a system with multiple right-hand sides, b should have size n by nrhs.

Keyword arguments

job::Symbol=:A: task to perform. Accepted values are

:A: solve Ax = b

:PL: solve PLx = Sb

:D: solve Dx = b

:LPS: solve L'P'S⁻¹x = b

:DLPS: solve DL'P'S⁻¹x = b.

Return values

x::Array{T}: an array of the same size as b containing the solutions.

Combined analysis, factorization and solve

ma97_solve(A, b; kwargs...)

Input arguments

A::SparseMatrixCSC{T<:Ma97Data,Int}: input matrix. A full matrix will be converted to sparse.

b::Array{T}: vector of array of right-hand sides. Note that b will be overwritten. To solve a system with multiple right-hand sides, b should have size n by nrhs.

Keyword arguments

matrix_type::Symbol=:real_indef: indicates the matrix type. Accepted values are

:real_spd for a real symmetric and positive definite matrix

:real_indef for a real symmetric and indefinite matrix.

Return values

x::Array{T}: an array of the same size as b containing the solutions.

HSL.ma97_solve! — Function.

In-place system solve

See the documentation for ma97_solve(). The only difference is that the right-hand side b is overwritten with the solution.